+++ ci/script.sh
@@ -491,7 +491,7 @@ main() {
             echo 'version = "0.3.0"' >> $td/Cargo.toml
 
             # Test MSP430
-            #test_svd_for_target msp430 https://raw.githubusercontent.com/pftbest/msp430g2553/v0.3.0-svd/msp430g2553.svd
+            test_svd_for_target msp430 https://raw.githubusercontent.com/pftbest/msp430g2553/v0.3.0-svd/msp430g2553.svd
         ;;
 
         # Community-provided RISC-V SVDs